\name{BOBgmms}
\alias{BOBgmms}
\docType{package}
\title{
\packageTitle{BOBgmms}
}
\description{
\packageDescription{BOBgmms}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BOBgmms}
\packageIndices{BOBgmms}
}
\author{
\packageAuthor{BOBgmms}

Maintainer: \packageMaintainer{BOBgmms}
}
\references{
  Marin, S., Loong, B., & Westveld, A. (2024). BOB: Bayesian Optimized Bootstrap
  for Uncertainty Quantification in Gaussian Mixture Models. arXiv preprint arXiv:2311.03644.
}
