% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOBgmms-package.R
\docType{package}
\name{BOBgmms}
\alias{BOBgmms-package}
\alias{BOBgmms}
\title{BOBgmms: Bayesian Optimized Bootstrap for Approximate Posterior Sampling in 
Gaussian Mixture Models}
\description{
Approximately sample from the posterior distribution of Gaussian mixture 
models under conjugate priors with the Weighted Bayesian Bootstrap (Newton 
et al. 2021 <https://doi.org/10.1002/cjs.11570>) and the Bayesian Optimized 
Bootstrap (Marin et al. 2024+ <https://doi.org/10.48550/arXiv.2311.03644>).
}
\references{
S. Marin, B. Long,and A. H. Westveld (2024), BOB: Bayesian Optimized 
Bootstrap for Approximate Posterior Sampling in Gaussian Mixture Models. 
\emph{arXiv}, {2311.03644}.

M.A. Newton, N.G. Polson, and J. Xu (2021), Weighted Bayesian bootstrap for 
scalable posterior distributions. \emph{Can J Statistics}, 49:421-437.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/marinsantiago/BOBgmms}
  \item Report bugs at \url{https://github.com/marinsantiago/BOBgmms/issues}
}

}
\author{
\strong{Maintainer}: Santiago Marin \email{santiago.marinardila@anu.edu.au} (\href{https://orcid.org/0009-0006-0086-3721}{ORCID})

}
