% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_densities.R
\name{loglik}
\alias{loglik}
\title{Log-likelihood Function}
\usage{
loglik(means, covs, probs, y)
}
\arguments{
\item{means}{A list of size \eqn{K} containing the mean vectors at which we 
want to evaluate the log-posterior density. Each vector of should be of
size \eqn{d}.}

\item{covs}{A list of of size \eqn{K} containing the covariance matrices at
which we want to evaluate the log-posterior density. Each matrix should be 
of size \eqn{d}-by-\eqn{d}.}

\item{probs}{A vector of size \eqn{d} of cluster probabilities at which we 
want to evaluate the log-posterior density. The sum of all the 
probabilities should equals one.}

\item{y}{A matrix of observations of dimension \eqn{n}-by-\eqn{d}, where each 
of the \eqn{n} rows is an observation vector.}
}
\value{
The evaluated log-likelihood as a scalar (numeric).
}
\description{
This function evaluates the log-likelihood of a Gaussian mixture model.
}
\author{
Santiago Marin
}
