% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_densities.R
\name{lpost}
\alias{lpost}
\title{Log-posterior Density}
\usage{
lpost(means, covs, probs, y, betas, lambdas, nus, psis, alphas)
}
\arguments{
\item{means}{A list of size \eqn{K} containing the mean vectors at which we 
want to evaluate the log-posterior density. Each vector of should be of
size \eqn{d}.}

\item{covs}{A list of of size \eqn{K} containing the covariance matrices at
which we want to evaluate the log-posterior density. Each matrix should be 
of size \eqn{d}-by-\eqn{d}.}

\item{probs}{A vector of size \eqn{d} of cluster probabilities at which we 
want to evaluate the log-posterior density. The sum of all the 
probabilities should equals one.}

\item{y}{A matrix of observations of dimension \eqn{n}-by-\eqn{d}, where each 
of the \eqn{n} rows is an observation vector.}

\item{betas}{A list of size \eqn{K} containing the prior means of each mean
vector. Each prior mean should be a vector of size \eqn{d}.}

\item{lambdas}{A vector of size \eqn{K} containing the \eqn{\lambda}
shrinkage hyper-parameters associated with each mean vector. Each element 
in the vector must be positive.}

\item{nus}{A vector of size \eqn{K} containing the \eqn{\nu} hyper-parameters 
associated with each covariance matrix. Each element in the vector must be 
larger than \eqn{d + 1}.}

\item{psis}{A list of size \eqn{K} containing the prior scale matrices of 
each covariance matrix. Each element in the list must be a symmetric
positive semidefinite matrix of size \eqn{d}-by-\eqn{d}.}

\item{alphas}{A vector of size \eqn{K} of Dirichlet prior hyper-parameters.
All the elements in the vector must be larger than one.}
}
\value{
The evaluated log-posterior density as a scalar (numeric).
}
\description{
This function evaluates the log-posterior density of a Gaussian mixture model
with conjugate priors.
}
\author{
Santiago Marin
}
