% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bob.R
\name{bob.gmm}
\alias{bob.gmm}
\title{Sample from the posterior distribution of a GMM via BOB}
\usage{
bob.gmm(
  y,
  means.init,
  covs.init,
  probs.init,
  betas,
  lambdas,
  nus,
  psis,
  alphas,
  lower_bound,
  upper_bound,
  max.iters = 20000,
  size.batch = 1000,
  bo.iters = 100,
  a = 0,
  b = 0,
  c = 1,
  r = 1,
  cores = parallel::detectCores() - 1,
  seed = sample.int(.Machine$integer.max, 1)
)
}
\arguments{
\item{y}{A matrix of observations of dimension \eqn{n}-by-\eqn{d}, where each 
of the \eqn{n} rows is an observation vector.}

\item{means.init}{A list of size \eqn{K} containing the initial mean 
vectors, each vector of size \eqn{d}.}

\item{covs.init}{A list of of size \eqn{K} containing the initial covariance 
matrices, each matrix of size \eqn{d}-by-\eqn{d}.}

\item{probs.init}{A vector of size \eqn{d} of initial probabilities. The sum
of all the probabilities should equals one.}

\item{betas}{A list of size \eqn{K} containing the prior means of each mean
vector. Each prior mean should be a vector of size \eqn{d}.}

\item{lambdas}{A vector of size \eqn{K} containing the \eqn{\lambda}
shrinkage hyper-parameters associated with each mean vector. Each element 
in the vector must be positive.}

\item{nus}{A vector of size \eqn{K} containing the \eqn{\nu} hyper-parameters 
associated with each covariance matrix. Each element in the vector must be 
larger than \eqn{d + 1}.}

\item{psis}{A list of size \eqn{K} containing the prior scale matrices of 
each covariance matrix. Each element in the list must be a symmetric
positive semidefinite matrix of size \eqn{d}-by-\eqn{d}.}

\item{alphas}{A vector of size \eqn{K} of Dirichlet prior hyper-parameters.
All the elements in the vector must be larger than one.}

\item{lower_bound}{A vector of size \eqn{3K + 1} containing the lower bounds 
of the search space for the Bayesian optimization procedure. All elements 
in the vector should be positive.}

\item{upper_bound}{A vector of size \eqn{3K + 1} containing the upper bounds 
of the search space for the Bayesian optimization procedure. All elements 
in the vector should be positive.}

\item{max.iters}{A positive integer, corresponding to the total number of 
approximate posterior draws. Default is 20000.}

\item{size.batch}{A positive integer, corresponding to the size of the batch
of approximate posterior draws used to estimate the KL divergence between
the approximate posterior and the true Bayesian posterior. Default is 1000.}

\item{bo.iters}{A positive integer, corresponding to the total number of 
iterations in the Bayesian optimization procedure. Default is 100.}

\item{a}{A scalar corresponding to the \eqn{a} hyper-parameter in the 
tempering profile. Note that \eqn{a\in [0,1)}. Default is 0.}

\item{b}{A scalar corresponding to the \eqn{b} hyper-parameter in the 
tempering profile. Note that \eqn{b\in \mathbb{R}}. Default is 0.}

\item{c}{A scalar corresponding to the \eqn{c} hyper-parameter in the 
tempering profile. Note that \eqn{c>0}. Default is 1.}

\item{r}{A scalar corresponding to the \eqn{r} hyper-parameter in the 
tempering profile. Note that \eqn{r>0}. Default is 1.}

\item{cores}{The number of CPU cores to use during the sampling process.
Default is \code{parallel::detectCores() - 1}.}

\item{seed}{The seed for random number generation. Default is
\code{sample.int(.Machine$integer.max, 1)}}
}
\value{
A list of size two. The first element is a matrix of size 
  \eqn{S}-by-\eqn{\text{dim}(\boldsymbol{\theta})} of approximate posterior 
  draws, where \eqn{S} is the total number of posterior draws and 
  \eqn{\text{dim}(\boldsymbol{\theta})} is the total number of parameters in 
  the model. The first \eqn{K \times d} columns in the matrix are posterior 
  draws from the mean parameters, the next \eqn{K\times d^2} columns are 
  posterior draws from variance and covariance parameters, and the last 
  \eqn{K} columns are posterior draws from the probability parameters. The 
  second elements in the list is a vector containing the optimal 
  \eqn{\boldsymbol{x}^{*}} values used in the random weighting. 
  See Marin et al. (2025+) for details.
}
\description{
This function generates approximate posterior draws from Gaussian mixture
models using the Bayesian Optimized Bootstrap (BOB, Marin et al. 2025+). The
function leverages parallel computing across multiple CPU cores with the 
package \code{parallel}. Please be aware that the parallelization is 
conducted via \emph{forking} rather than \emph{sockets}, so it only works on 
"unix" systems. One can verify the OS type by running the following R code:
\code{.Platform$OS.type}.
}
\references{
S. Marin, B. Long,and A. H. Westveld (2025+), BOB: Bayesian Optimized 
Bootstrap for Approximate Posterior Sampling in Gaussian Mixture Models. 
\emph{arXiv}, {2311.03644}.
}
\author{
Santiago Marin
}
