% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOBgmms-package.R
\docType{package}
\name{BOBgmms}
\alias{BOBgmms-package}
\alias{BOBgmms}
\title{BOBgmms: Bayesian optimized bootstrap for approximate posterior sampling in 
Gaussian mixture models}
\description{
Approximately sample from the posterior distribution of Gaussian mixture 
models under conjugate priors with the weighted Bayesian bootstrap (Newton 
et al. 2020 <https://doi.org/10.1002/cjs.11570>) and the Bayesian optimized 
bootstrap (Marin et al. 2026 <https://doi.org/10.1007/s11222-025-10763-y>).
}
\references{
S. Marin, B. Long,and A. H. Westveld (2026), BOB: Bayesian optimized 
bootstrap for approximate posterior sampling in Gaussian mixture models. 
\emph{Statistics and Computing}, 36, 14.

M.A. Newton, N.G. Polson, and J. Xu (2021), Weighted Bayesian bootstrap for 
scalable posterior distributions. \emph{Can J Statistics}, 49:421-437.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/marinsantiago/BOBgmms}
  \item Report bugs at \url{https://github.com/marinsantiago/BOBgmms/issues}
}

}
\author{
\strong{Maintainer}: Santiago Marin \email{santiago.marinardila@anu.edu.au} (\href{https://orcid.org/0009-0006-0086-3721}{ORCID})

Other contributors:
\itemize{
  \item Bronwyn Loong \email{bronwyn.loong@anu.edu.au} (\href{https://orcid.org/0000-0002-1409-8892}{ORCID}) [contributor]
  \item Anton H. Westveld \email{anton.westveld@anu.edu.au} (\href{https://orcid.org/0000-0003-1671-0931}{ORCID}) [contributor]
}

}
