% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_optim.R
\name{em.optim}
\alias{em.optim}
\title{Log-posterior optimization via expectation-maximization (EM).}
\usage{
em.optim(
  y,
  means.init,
  covs.init,
  probs.init,
  betas,
  lambdas,
  nus,
  psis,
  alphas,
  lik.weights = rep(1, nrow(y)),
  means.weights = rep(1, length(means.init)),
  covs.weights = rep(1, length(means.init)),
  probs.weights = rep(1, length(means.init)),
  a = 0,
  b = 0,
  c = 1,
  r = 1,
  max.iters = 500,
  epss = 1e-04
)
}
\arguments{
\item{y}{A matrix of observations of dimension \eqn{n}-by-\eqn{d}, where each 
of the \eqn{n} rows is an observation vector.}

\item{means.init}{A list of size \eqn{K} containing the initial mean 
vectors, each vector of size \eqn{d}.}

\item{covs.init}{A list of of size \eqn{K} containing the initial covariance 
matrices, each matrix of size \eqn{d}-by-\eqn{d}.}

\item{probs.init}{A vector of size \eqn{d} of initial probabilities. The sum
of all the probabilities should equals one.}

\item{betas}{A list of size \eqn{K} containing the prior means of each mean
vector. Each prior mean should be a vector of size \eqn{d}.}

\item{lambdas}{A vector of size \eqn{K} containing the \eqn{\lambda}
shrinkage hyper-parameters associated with each mean vector. Each element 
in the vector must be positive.}

\item{nus}{A vector of size \eqn{K} containing the \eqn{\nu} hyper-parameters 
associated with each covariance matrix. Each element in the vector must be 
larger than \eqn{d + 1}.}

\item{psis}{A list of size \eqn{K} containing the prior scale matrices of 
each covariance matrix. Each element in the list must be a symmetric
positive semidefinite matrix of size \eqn{d}-by-\eqn{d}.}

\item{alphas}{A vector of size \eqn{K} of Dirichlet prior hyper-parameters.
All the elements in the vector must be larger than one.}

\item{lik.weights}{A vector of size \eqn{n} of likelihood weights. All 
weights must be non-negative. Default is 1 for each weight, i.e., no 
weighting.}

\item{means.weights}{A vector of size \eqn{K} of weights associated with each
mean vector. All weights must be non-negative. Default is 1 for each 
weight, i.e., no weighting.}

\item{covs.weights}{A vector of size \eqn{K} of weights associated with each
covariance matrix. All weights must be non-negative. Default is 1 for each 
weight, i.e., no weighting.}

\item{probs.weights}{A vector of size \eqn{K} of weights associated with each
probability. All weights must be non-negative. Default is 1 for each 
weight, i.e., no weighting.}

\item{a}{A scalar corresponding to the \eqn{a} hyper-parameter in the 
tempering profile. Constraint: \eqn{a\in [0,1)}. Default is 0.}

\item{b}{A scalar corresponding to the \eqn{b} hyper-parameter in the 
tempering profile. Constraint: \eqn{b\in \mathbb{R}}. Default is 0.}

\item{c}{A scalar corresponding to the \eqn{c} hyper-parameter in the 
tempering profile. Constraint: \eqn{c>0}. Default is 1.}

\item{r}{A scalar corresponding to the \eqn{r} hyper-parameter in the 
tempering profile. Constraint: \eqn{r>0}. Default is 1.}

\item{max.iters}{Maximum number of EM iterations. Default is 500.}

\item{epss}{Convergence threshold for the EM algorithm. Default is 
\code{1e-4}}
}
\value{
A list containing the EM solution for each mean vector, each 
  covariance matrix, and each mixture component.
}
\description{
This function maximizes the weighted log-posterior density associated with a 
Gaussian mixture model with the EM algorithm as described in Marin et 
al. (2026).
}
\references{
S. Marin, B. Long,and A. H. Westveld (2026), BOB: Bayesian optimized 
bootstrap for approximate posterior sampling in Gaussian mixture models. 
\emph{Statistics and Computing}, 36, 14.
}
\author{
Santiago Marin
}
