% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbb.R
\name{wbb.gmm}
\alias{wbb.gmm}
\title{Approximately sample from the posterior distribution of a GMM via WBB}
\usage{
wbb.gmm(
  y,
  means.init,
  covs.init,
  probs.init,
  betas,
  lambdas,
  nus,
  psis,
  alphas,
  wbb.scheme = "wbb1",
  max.iters = 20000,
  a = 0,
  b = 0,
  c = 1,
  r = 1,
  cores = parallel::detectCores() - 1,
  seed = sample.int(.Machine$integer.max, 1)
)
}
\arguments{
\item{y}{A matrix of observations of dimension \eqn{n}-by-\eqn{d}, where each 
of the \eqn{n} rows is an observation vector.}

\item{means.init}{A list of size \eqn{K} containing the initial mean 
vectors, each vector of size \eqn{d}.}

\item{covs.init}{A list of of size \eqn{K} containing the initial covariance 
matrices, each matrix of size \eqn{d}-by-\eqn{d}.}

\item{probs.init}{A vector of size \eqn{d} of initial probabilities. The sum
of all the probabilities should equals one.}

\item{betas}{A list of size \eqn{K} containing the prior means of each mean
vector. Each prior mean should be a vector of size \eqn{d}.}

\item{lambdas}{A vector of size \eqn{K} containing the \eqn{\lambda}
shrinkage hyper-parameters associated with each mean vector. Each element 
in the vector must be positive.}

\item{nus}{A vector of size \eqn{K} containing the \eqn{\nu} hyper-parameters 
associated with each covariance matrix. Each element in the vector must be 
larger than \eqn{d + 1}.}

\item{psis}{A list of size \eqn{K} containing the prior scale matrices of 
each covariance matrix. Each element in the list must be a symmetric
positive semidefinite matrix of size \eqn{d}-by-\eqn{d}.}

\item{alphas}{A vector of size \eqn{K} of Dirichlet prior hyper-parameters.
All the elements in the vector must be larger than one.}

\item{wbb.scheme}{A character string denoting which weighting scheme to use.
The options are either (i) \code{"wbb1"} for random prior weights or (ii) 
\code{"wbb2"} for fixed prior weights. Default is \code{"wbb1"}.}

\item{max.iters}{A positive integer, corresponding to the total number of 
approximate posterior draws. Default is 20000.}

\item{a}{A scalar corresponding to the \eqn{a} hyper-parameter in the 
tempering profile. Constraint: \eqn{a\in [0,1)}. Default is 0.}

\item{b}{A scalar corresponding to the \eqn{b} hyper-parameter in the 
tempering profile. Constraint: \eqn{b\in \mathbb{R}}. Default is 0.}

\item{c}{A scalar corresponding to the \eqn{c} hyper-parameter in the 
tempering profile. Constraint: \eqn{c>0}. Default is 1.}

\item{r}{A scalar corresponding to the \eqn{r} hyper-parameter in the 
tempering profile. Constraint: \eqn{r>0}. Default is 1.}

\item{cores}{The number of CPU cores to use during the sampling process.
Default is \code{parallel::detectCores() - 1}.}

\item{seed}{The seed for random number generation. Default is
\code{sample.int(.Machine$integer.max, 1)}}
}
\value{
A matrix of size \eqn{S}-by-\eqn{\text{dim}(\boldsymbol{\theta})}
  of approximate posterior draws, where \eqn{S} is the total number of
  posterior draws and \eqn{\text{dim}(\boldsymbol{\theta})} is the total
  number of parameters in the model. The first \eqn{K \times d} columns are 
  posterior draws from the mean parameters, the next \eqn{K\times d^2} 
  columns are posterior draws from variance and covariance parameters, and 
  the last \eqn{K} columns are posterior draws from the probability 
  parameters.
}
\description{
This function generates approximate draws for the posterior distribution of a
Gaussian mixture model under conjugate prior using the weighted Bayesian 
bootstrap (WBB, Newton et al. 2021). The function supports two options of 
prior weights: (i) Random prior weights or (ii) fixed prior weights. See 
Newton et al. (2021) for additional details.
}
\details{
\emph{Note}: The function leverages parallel computing across multiple CPU 
cores through the package \code{parallel}. Please be aware that 
parallelization over multiple CPU workers is conducted via \emph{forking} 
rather than \emph{sockets}, so it is only available on POSIX ("unix") 
systems. On non-POSIX platforms, the function is still operational, but the 
number of CPU workers will be automatically set to one. To verify your 
operating system (OS), simply run the following R code: 
\code{.Platform$OS.type}.
}
\references{
M.A. Newton, N.G. Polson, and J. Xu (2021), Weighted Bayesian bootstrap for 
scalable posterior distributions. \emph{Can J Statistics}, 49:421-437.
}
\author{
Santiago Marin
}
